Separate Azure CI logs
During CI, log each target to its own file so that that file can be
displayed separately later. This makes the Prow logs look cleaner and
easier to parse.

To keep target list the same between the Makefile and the CI script,
move the Azure targets into a common azure_targets.sh file, and
source/parse this file in both the Makefile and the bash script.