Fix OVA CI script to catch background errors
The OVA CI script was using `wait` with no arguments to wait until all
background processes finished before proceeding. From `help wait` in
BASH:

  - If ID is not given, waits for all currently active child processes,
    and the return status is zero.

I know I've seen the script fail when a background `make` failed,
so I am confused on why that would happen if `wait` was always exiting
with a 0 return code. Regardless, I ran experiments where I forced the
first `make` target to fail, and saw that the script exited with 0,
showing that tests passed when they didnt. The observed behavior was
`wait` returning the exit code of the last checked child process.

This patch changes the behavior to explicitly check the return code
for each child process, and exits non-zero if any of them are
non-zero.