photon: disable apparmor in kernel
Previous attempts to disable AppArmor on Photon were not sufficient. So
long as AppArmor is enabled in the kernel, kubelet and containerD are
loading an apparmor profile that gets enforced and does not work on
Photon at this time.

This patch appends `apparmor=0` to the kernel boot command line to
disable.