capi: make PATH extension work when ran from a different directory
When `make` is executed as `make -C images/cap`, PWD does not point to
images/capi, but to the directory where the user is located. Thus extending
PATH to find the packer fetched in 'ensure-packer.sh' was not working. Fix this
by using CURDIR instead of PWD, which will work in both cases.