Clean up Flatcar provisioner
Since we already check the build name inside execute_command, there
is no need to check the same again inside bootstrap-flatcar.sh.
Moreover, having both conditionals is confusing and error-prone.

Putting the conditional in the Packer config is more visible than
doing so inside the script. It's arguably also more intuitive, since
now bootstrap-flatcar.sh is only executed when the build is a Flatcar
build (as opposed to always executing and then exiting when the build
isn't Flatcar).

So, we keep the conditional in execute_command and remove the one in
bootstrap-flatcar.sh.

- Remove the conditional inside bootstrap-flatcar.sh.
- Remove environment_vars since it affects only bootstrap-flatcar.sh
  and has no influence on execute_command.

Signed-off-by: Johanan Liebermann <jliebermann@microsoft.com>